+++ src/Graph/ShortestPathAlgo.java
@@ -1,9 +1,7 @@
 package Graph;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.LinkedList;
-import java.util.Queue;
+import java.util.*;
+import java.util.stream.Collectors;
 
 public class ShortestPathAlgo {
     // TC = O(E + V)
@@ -80,5 +78,54 @@ public class ShortestPathAlgo {
     private static void addEdge(ArrayList<ArrayList<Integer>> adj, int vertex1, int vertex2) {
         adj.get(vertex1).add(vertex2);
         adj.get(vertex2).add(vertex1);
+    }
+    /*
+    Input: n = 5, roads = [[0,1],[1,2],[2,3],[0,2],[1,3],[2,4]]
+    Output: 43
+    Explanation: The figure above shows the country and the assigned values of [2,4,5,3,1].
+    - The road (0,1) has an importance of 2 + 4 = 6.
+    - The road (1,2) has an importance of 4 + 5 = 9.
+    - The road (2,3) has an importance of 5 + 3 = 8.
+    - The road (0,2) has an importance of 2 + 5 = 7.
+    - The road (1,3) has an importance of 4 + 3 = 7.
+    - The road (2,4) has an importance of 5 + 1 = 6.
+    The total importance of all roads is 6 + 9 + 8 + 7 + 7 + 6 = 43.
+    It can be shown that we cannot obtain a greater total importance than 43.=
+    */
+    //Author: Anand
+    // TC = O(nlogn)
+    public long maximumImportance(int n, int[][] roads) {
+        Map<Integer, Integer> nodesConnectedCount = new HashMap<>();
+        for (int[] r : roads) {
+            nodesConnectedCount.put(r[0], nodesConnectedCount.getOrDefault(r[0], 0) + 1);
+            nodesConnectedCount.put(r[1], nodesConnectedCount.getOrDefault(r[1], 0) + 1);
+        }
+
+
+        nodesConnectedCount = sortByValueInteger(nodesConnectedCount);
+
+        Map<Integer, Integer> assign = new HashMap<>();
+        int max = n;
+        for (Map.Entry<Integer, Integer> entry : nodesConnectedCount.entrySet()) assign.put((int) entry.getKey(), n--);
+
+        long ans = 0L;
+        for (int[] r : roads) ans += assign.get(r[0]) + assign.get(r[1]);
+
+        return ans;
+
+    }
+
+    // function to sort hashmap by values
+    public static HashMap<Integer, Integer> sortByValueInteger(Map<Integer, Integer> hm) {
+        HashMap<Integer, Integer> temp
+                = hm.entrySet()
+                .stream()
+                .sorted((i1, i2) -> i2.getValue().compareTo(i1.getValue()))
+                .collect(Collectors.toMap(
+                        Map.Entry::getKey,
+                        Map.Entry::getValue,
+                        (e1, e2) -> e1, LinkedHashMap::new));
+
+        return temp;
     }
 }
