+++ src/com/company/DynamicArray.java
@@ -9864,94 +9864,94 @@ class SubSet2 {
     }
 
 
-class Subsets {
-    // TC = O(n), SC = O(n)
+    class Subsets {
+        // TC = O(n), SC = O(n)
 
-    public List<List<Integer>> subsets(int[] nums) {
-        int n = nums.length;
-        LinkedList<List<Integer>> result = new LinkedList<>();
-        if (n == 0) return result;
-        result.add(new ArrayList<>());
-        permuate(nums, 0, result);
-        return result;
-    }
-
-    private void permuate(int[] nums, int index, LinkedList<List<Integer>> result) {
-        int n = nums.length;
-        // base case
-        if (index == n) {
-            return;
+        public List<List<Integer>> subsets(int[] nums) {
+            int n = nums.length;
+            LinkedList<List<Integer>> result = new LinkedList<>();
+            if (n == 0) return result;
+            result.add(new ArrayList<>());
+            permuate(nums, 0, result);
+            return result;
         }
 
-        int ch = nums[index];
-        int size = result.size();
-        LinkedList<Set<Integer>> list = new LinkedList<Set<Integer>>();
+        private void permuate(int[] nums, int index, LinkedList<List<Integer>> result) {
+            int n = nums.length;
+            // base case
+            if (index == n) {
+                return;
+            }
 
-        while (size-- > 0) {
-            List<Integer> elem = result.poll();
-            list.add(new HashSet<>(elem));
-            elem.add(ch);
-            list.add(new HashSet<>(elem));
+            int ch = nums[index];
+            int size = result.size();
+            LinkedList<Set<Integer>> list = new LinkedList<Set<Integer>>();
+
+            while (size-- > 0) {
+                List<Integer> elem = result.poll();
+                list.add(new HashSet<>(elem));
+                elem.add(ch);
+                list.add(new HashSet<>(elem));
+            }
+            result.addAll(list.stream().map(x -> new ArrayList<>(x)).collect(Collectors.toList()));
+
+            permuate(nums, index + 1, result);
         }
-        result.addAll(list.stream().map(x -> new ArrayList<>(x)).collect(Collectors.toList()));
-
-        permuate(nums, index + 1, result);
     }
-}
 
 // Iterative approach
 
 // Algorithm :-
 
-/*
- - Idea is to take all paths by variation of a  single character (ie. uppercase or lowercase)
- - Intialise list with first given variation of string
- - Iterate two loops
-   - i = 0 unitll n (n = size of string)
-   - check for current char ie. if its character then its liable for changes else not
-   - poll elements out of list untill it becomes empty
-   - create both variation and update the stirng to list
-   - Repeat the process untill while string gets exhausted
-  - Return all possible paths
-*/
+    /*
+     - Idea is to take all paths by variation of a  single character (ie. uppercase or lowercase)
+     - Intialise list with first given variation of string
+     - Iterate two loops
+       - i = 0 unitll n (n = size of string)
+       - check for current char ie. if its character then its liable for changes else not
+       - poll elements out of list untill it becomes empty
+       - create both variation and update the stirng to list
+       - Repeat the process untill while string gets exhausted
+      - Return all possible paths
+    */
 // Recursive approach
-class LetterCasePermutation {
-    // TC = O(2*n), SC = O(n!)
-    public List<String> letterCasePermutation(String s) {
-        int n = s.length();
-        List<String> ans = new ArrayList<>();
-        if (n == 0) return ans;
-        StringBuilder sb = new StringBuilder("");
-        permuate(s, sb, 0, ans);
-        return ans;
-    }
-
-    private void permuate(String s, StringBuilder sb, int index, List<String> ans) {
-        int n = s.length();
-        // base case
-        if (index == n) {
-            ans.add(sb.toString());
-            return;
+    class LetterCasePermutation {
+        // TC = O(2*n), SC = O(n!)
+        public List<String> letterCasePermutation(String s) {
+            int n = s.length();
+            List<String> ans = new ArrayList<>();
+            if (n == 0) return ans;
+            StringBuilder sb = new StringBuilder("");
+            permuate(s, sb, 0, ans);
+            return ans;
         }
 
-        char c = s.charAt(index);
+        private void permuate(String s, StringBuilder sb, int index, List<String> ans) {
+            int n = s.length();
+            // base case
+            if (index == n) {
+                ans.add(sb.toString());
+                return;
+            }
 
-        // if char is letter
-        if (Character.isLetter(c)) {
-            Character[] array = new Character[]{Character.toUpperCase(c), Character.toLowerCase(c)};
-            for (char ch : array) {
-                sb.append(ch);
+            char c = s.charAt(index);
+
+            // if char is letter
+            if (Character.isLetter(c)) {
+                Character[] array = new Character[]{Character.toUpperCase(c), Character.toLowerCase(c)};
+                for (char ch : array) {
+                    sb.append(ch);
+                    permuate(s, sb, index + 1, ans);
+                    sb.setLength(sb.length() - 1);
+                }
+            }
+            // else do not make any variation
+            else {
+                sb.append(c);
                 permuate(s, sb, index + 1, ans);
                 sb.setLength(sb.length() - 1);
             }
         }
-        // else do not make any variation
-        else {
-            sb.append(c);
-            permuate(s, sb, index + 1, ans);
-            sb.setLength(sb.length() - 1);
-        }
-    }
 
     /*
 
@@ -9984,159 +9984,159 @@ class LetterCasePermutation {
         return new ArrayList<>(result);
     }
      */
-}
+    }
 
-class PickRandomBlacklist {
-    HashMap<Integer, Integer> map;
-    Random generator;
-    int ul;
+    class PickRandomBlacklist {
+        HashMap<Integer, Integer> map;
+        Random generator;
+        int ul;
 
-    public PickRandomBlacklist(int n, int[] blacklist) {
-        this.generator = new Random();
-        this.map = new HashMap<>();
-        for (Integer num : blacklist)
-            this.map.put(num, -1);
+        public PickRandomBlacklist(int n, int[] blacklist) {
+            this.generator = new Random();
+            this.map = new HashMap<>();
+            for (Integer num : blacklist)
+                this.map.put(num, -1);
 
-        this.ul = n - blacklist.length;
-        int j = n - 1;
-        for (Integer num : blacklist) {
-            if (num < ul) {
-                while (map.containsKey(j)) j--;
-                map.put(num, j);
-                j--;
+            this.ul = n - blacklist.length;
+            int j = n - 1;
+            for (Integer num : blacklist) {
+                if (num < ul) {
+                    while (map.containsKey(j)) j--;
+                    map.put(num, j);
+                    j--;
+                }
             }
+        }
+
+        public int pick() {
+            int random = generator.nextInt(ul);
+            if (map.containsKey(random))
+                return map.get(random);
+            return random;
         }
     }
 
-    public int pick() {
-        int random = generator.nextInt(ul);
-        if (map.containsKey(random))
-            return map.get(random);
-        return random;
-    }
-}
+    /**
+     * Your Solution object will be instantiated and called as such:
+     * Solution obj = new Solution(n, blacklist);
+     * int param_1 = obj.pick();
+     */
 
-/**
- * Your Solution object will be instantiated and called as such:
- * Solution obj = new Solution(n, blacklist);
- * int param_1 = obj.pick();
- */
+    class FindDifferentBinaryString {
+        Set<String> set;
+        String res;
+        int len;
+        String[] d = {
+                "1", "0"
+        };
 
-class FindDifferentBinaryString {
-    Set<String> set;
-    String res;
-    int len;
-    String[] d = {
-            "1", "0"
-    };
+        public String findDifferentBinaryString(String[] nums) {
+            len = nums.length;
+            set = new HashSet<>();
+            for (String s : nums) set.add(s);
+            diff(0, "");
+            return res;
+        }
 
-    public String findDifferentBinaryString(String[] nums) {
-        len = nums.length;
-        set = new HashSet<>();
-        for (String s : nums) set.add(s);
-        diff(0, "");
-        return res;
-    }
-
-    private Boolean diff(int index, String str) {
-        // base case
-        if (index == set.size()) {
-            if (!set.contains(str)) {
-                res = str;
-                return true;
+        private Boolean diff(int index, String str) {
+            // base case
+            if (index == set.size()) {
+                if (!set.contains(str)) {
+                    res = str;
+                    return true;
+                }
+                return false;
+            }
+            // For all position place string characters
+            for (int k = 0; k < len; k++) {
+                for (int i = 0; i < d.length; i++) {
+                    str += d[i];
+                    if (diff(index + 1, str)) return true;
+                    str = str.substring(0, str.length() - 1);
+                }
             }
             return false;
         }
-        // For all position place string characters
-        for (int k = 0; k < len; k++) {
-            for (int i = 0; i < d.length; i++) {
-                str += d[i];
-                if (diff(index + 1, str)) return true;
-                str = str.substring(0, str.length() - 1);
+    }
+
+    /**
+     * Definition for singly-linked list.
+     * public class ListNode {
+     * int val;
+     * ListNode next;
+     * ListNode() {}
+     * ListNode(int val) { this.val = val; }
+     * ListNode(int val, ListNode next) { this.val = val; this.next = next; }
+     * }
+     */
+    class ReverseLinkedList {
+        public ListNode reverseList(ListNode head) {
+
+            if (head == null) return head;
+            ListNode curr = head, prev = null, next;
+            while (curr != null) {
+                next = curr.next;
+                curr.next = prev;
+                prev = curr;
+                curr = next;
             }
+            head = prev;
+            return head;
         }
-        return false;
-    }
-}
-
-/**
- * Definition for singly-linked list.
- * public class ListNode {
- * int val;
- * ListNode next;
- * ListNode() {}
- * ListNode(int val) { this.val = val; }
- * ListNode(int val, ListNode next) { this.val = val; this.next = next; }
- * }
- */
-class ReverseLinkedList {
-    public ListNode reverseList(ListNode head) {
-
-        if (head == null) return head;
-        ListNode curr = head, prev = null, next;
-        while (curr != null) {
-            next = curr.next;
-            curr.next = prev;
-            prev = curr;
-            curr = next;
-        }
-        head = prev;
-        return head;
-    }
-}
-
-/**
- * Definition for singly-linked list.
- * public class ListNode {
- * int val;
- * ListNode next;
- * ListNode() {}
- * ListNode(int val) { this.val = val; }
- * ListNode(int val, ListNode next) { this.val = val; this.next = next; }
- * }
- */
-class ReverseLinkedListKGroup {
-    public ListNode reverseKGroup(ListNode head, int k) {
-        if (head == null) return head;
-        int len = getLength(head);
-        return reverse(head, k, len);
     }
 
-    private int getLength(ListNode curr) {
-        int cnt = 0;
-        while (curr != null) {
-            cnt++;
-            curr = curr.next;
-        }
-        return cnt;
-    }
-
-    // Recursive function to reverse in groups
-    private ListNode reverse(ListNode head, int k, int len) {
-        // base case
-        if (len < k) return head;
-
-        ListNode curr = head, prev = null, next = null;
-        for (int i = 0; i < k; i++) {
-            // reverse linked list code
-            next = curr.next;
-            curr.next = prev;
-            prev = curr;
-            curr = next;
+    /**
+     * Definition for singly-linked list.
+     * public class ListNode {
+     * int val;
+     * ListNode next;
+     * ListNode() {}
+     * ListNode(int val) { this.val = val; }
+     * ListNode(int val, ListNode next) { this.val = val; this.next = next; }
+     * }
+     */
+    class ReverseLinkedListKGroup {
+        public ListNode reverseKGroup(ListNode head, int k) {
+            if (head == null) return head;
+            int len = getLength(head);
+            return reverse(head, k, len);
         }
 
-        ListNode nextNode = reverse(curr, k, len - k);
-        head.next = nextNode;
-        return prev;
+        private int getLength(ListNode curr) {
+            int cnt = 0;
+            while (curr != null) {
+                cnt++;
+                curr = curr.next;
+            }
+            return cnt;
+        }
+
+        // Recursive function to reverse in groups
+        private ListNode reverse(ListNode head, int k, int len) {
+            // base case
+            if (len < k) return head;
+
+            ListNode curr = head, prev = null, next = null;
+            for (int i = 0; i < k; i++) {
+                // reverse linked list code
+                next = curr.next;
+                curr.next = prev;
+                prev = curr;
+                curr = next;
+            }
+
+            ListNode nextNode = reverse(curr, k, len - k);
+            head.next = nextNode;
+            return prev;
+        }
     }
-}
 
-class MaxTaxiEarningsDP {
+    class MaxTaxiEarningsDP {
 
-    // DP + Iteration based solution
-    public long maxTaxiEarnings(int n, int[][] rides) {
-        // sort rides  matrix based on  start time
-        Arrays.sort(rides, (a, b) -> a[0] - b[0]);
+        // DP + Iteration based solution
+        public long maxTaxiEarnings(int n, int[][] rides) {
+            // sort rides  matrix based on  start time
+            Arrays.sort(rides, (a, b) -> a[0] - b[0]);
 		/*
 		   [[2,5,4],[1,5,1]]
 		   sort ------> [[1 5 1], [2 5 4]]
@@ -10148,115 +10148,115 @@ class MaxTaxiEarningsDP {
          - Return max value ie. dp[n]
 		   Complexity -->  TC = O(n+k) , SC = O(n);
 		*/
-        long[] dp = new long[n + 1];
-        int j = 0;
-        for (int i = 1; i <= n; i++) {
-            dp[i] = Math.max(dp[i - 1], dp[i]);
-            while (j < rides.length && rides[j][0] == i) {
-                int d = rides[j][1] - rides[j][0] + rides[j][2];
-                dp[rides[j][1]] = Math.max(dp[rides[j][1]], dp[i] + d);
-                ++j;
+            long[] dp = new long[n + 1];
+            int j = 0;
+            for (int i = 1; i <= n; i++) {
+                dp[i] = Math.max(dp[i - 1], dp[i]);
+                while (j < rides.length && rides[j][0] == i) {
+                    int d = rides[j][1] - rides[j][0] + rides[j][2];
+                    dp[rides[j][1]] = Math.max(dp[rides[j][1]], dp[i] + d);
+                    ++j;
+                }
             }
+            return dp[n];
         }
-        return dp[n];
     }
-}
 
 
-class FindOriginalArray {
-    public int[] findOriginalArray(int[] changed) {
-        int n = changed.length, cnt = 0;
-        List<Integer> ans = new ArrayList<>();
-        HashMap<Integer, List<Integer>> map = new HashMap<>();
-        Arrays.sort(changed);
+    class FindOriginalArray {
+        public int[] findOriginalArray(int[] changed) {
+            int n = changed.length, cnt = 0;
+            List<Integer> ans = new ArrayList<>();
+            HashMap<Integer, List<Integer>> map = new HashMap<>();
+            Arrays.sort(changed);
 
-        // for (Integer ll: changed) System.out.println(ll);
+            // for (Integer ll: changed) System.out.println(ll);
 
-        if (n == 1 || n == 0 || n % 2 != 0) return new int[]{};
+            if (n == 1 || n == 0 || n % 2 != 0) return new int[]{};
 
-        for (int i = 0; i < n; i++) {
-            int e = changed[i];
-            if (e % 2 == 0 && map.containsKey(e / 2) && map.get(e / 2).size() > 0) {
-                ans.add(e / 2);
-                cnt++;
-                List<Integer> list = map.get(e / 2);
-                list.remove(0);
-                if (list.size() > 0) map.put(e / 2, list);
-                else map.remove(e / 2);
-                continue;
+            for (int i = 0; i < n; i++) {
+                int e = changed[i];
+                if (e % 2 == 0 && map.containsKey(e / 2) && map.get(e / 2).size() > 0) {
+                    ans.add(e / 2);
+                    cnt++;
+                    List<Integer> list = map.get(e / 2);
+                    list.remove(0);
+                    if (list.size() > 0) map.put(e / 2, list);
+                    else map.remove(e / 2);
+                    continue;
+                }
+                if (map.containsKey(e)) {
+                    List<Integer> l = map.get(e);
+                    l.add(i);
+                    map.put(e, l);
+                } else {
+                    List<Integer> l = new ArrayList<>();
+                    l.add(i);
+                    map.put(e, l);
+                }
             }
-            if (map.containsKey(e)) {
-                List<Integer> l = map.get(e);
-                l.add(i);
-                map.put(e, l);
+            if (cnt == n / 2) return ans.stream().mapToInt(x -> x).toArray();
+            return new int[]{};
+        }
+    }
+
+    // TODO: Solve again
+    class SplitIntoFibonacci {
+        List<Integer> al = new ArrayList<>();
+
+        public List<Integer> splitIntoFibonacci(String num) {
+            for (int i = 0; i < num.length(); i++) {
+                if (i > 0 && num.charAt(0) == '0') return new ArrayList<Integer>();
+                long a = Long.parseLong(num.substring(0, i + 1));
+                if (a > Integer.MAX_VALUE) break;
+                for (int j = i + 1; j < num.length(); j++) {
+                    long b = Long.parseLong(num.substring(i + 1, j + 1));
+                    if (b > Integer.MAX_VALUE) break;
+                    al.add((int) a);
+                    al.add((int) b);
+                    if (j > i + 1 && num.charAt(i + 1) == '0') break;
+                    if (j + 1 < num.length() && solve((int) a, (int) b, num, j + 1)) return al;
+                    else al.remove(al.size() - 1);
+                }
+                al.clear();
+            }
+            return new ArrayList<Integer>();
+        }
+
+        public boolean solve(int a, int b, String num, int index) {
+            if (index == num.length()) return true;
+            String sum = String.valueOf(a + b);
+            if (index + sum.length() > num.length()) {
+                al.remove(al.size() - 1);
+                return false;
+            }
+            String actualSum = num.substring(index, index + sum.length());
+            if (!sum.equals(actualSum)) {
+                al.remove(al.size() - 1);
+                return false;
             } else {
-                List<Integer> l = new ArrayList<>();
-                l.add(i);
-                map.put(e, l);
+                al.add(a + b);
+                return solve(b, a + b, num, index + sum.length());
             }
         }
-        if (cnt == n / 2) return ans.stream().mapToInt(x -> x).toArray();
-        return new int[]{};
-    }
-}
-
-// TODO: Solve again
-class SplitIntoFibonacci {
-    List<Integer> al = new ArrayList<>();
-
-    public List<Integer> splitIntoFibonacci(String num) {
-        for (int i = 0; i < num.length(); i++) {
-            if (i > 0 && num.charAt(0) == '0') return new ArrayList<Integer>();
-            long a = Long.parseLong(num.substring(0, i + 1));
-            if (a > Integer.MAX_VALUE) break;
-            for (int j = i + 1; j < num.length(); j++) {
-                long b = Long.parseLong(num.substring(i + 1, j + 1));
-                if (b > Integer.MAX_VALUE) break;
-                al.add((int) a);
-                al.add((int) b);
-                if (j > i + 1 && num.charAt(i + 1) == '0') break;
-                if (j + 1 < num.length() && solve((int) a, (int) b, num, j + 1)) return al;
-                else al.remove(al.size() - 1);
-            }
-            al.clear();
-        }
-        return new ArrayList<Integer>();
     }
 
-    public boolean solve(int a, int b, String num, int index) {
-        if (index == num.length()) return true;
-        String sum = String.valueOf(a + b);
-        if (index + sum.length() > num.length()) {
-            al.remove(al.size() - 1);
-            return false;
-        }
-        String actualSum = num.substring(index, index + sum.length());
-        if (!sum.equals(actualSum)) {
-            al.remove(al.size() - 1);
-            return false;
-        } else {
-            al.add(a + b);
-            return solve(b, a + b, num, index + sum.length());
-        }
-    }
-}
 
-
-/**
- * Definition for a binary tree node.
- * public class TreeNode {
- * int val;
- * TreeNode left;
- * TreeNode right;
- * TreeNode() {}
- * TreeNode(int val) { this.val = val; }
- * TreeNode(int val, TreeNode left, TreeNode right) {
- * this.val = val;
- * this.left = left;
- * this.right = right;
- * }
- * }
- */
+    /**
+     * Definition for a binary tree node.
+     * public class TreeNode {
+     * int val;
+     * TreeNode left;
+     * TreeNode right;
+     * TreeNode() {}
+     * TreeNode(int val) { this.val = val; }
+     * TreeNode(int val, TreeNode left, TreeNode right) {
+     * this.val = val;
+     * this.left = left;
+     * this.right = right;
+     * }
+     * }
+     */
 // Algo:-
 
 // - if tree is null retrun empty array list
@@ -10273,527 +10273,527 @@ TC = O(Nlogn) + O(N) ~= O(nlogn)
 SC = O(N)
 */
 
-class VerticalTraversal {
-    public class Pair implements Comparable<Pair> {
-        int key;
-        int value;
+    class VerticalTraversal {
+        public class Pair implements Comparable<Pair> {
+            int key;
+            int value;
 
-        public Pair(int key, int value) //Constructor of the class
-        {
-            this.key = key;
-            this.value = value;
-        }
-
-        public int compareTo(Pair o) {
-            return this.key - o.key;
-        }
-    }
-
-    // To store the vline and nodes with level on that vline
-    HashMap<Integer, PriorityQueue<Pair>> map = new HashMap<>();
-
-    public List<List<Integer>> verticalTraversal(TreeNode root) {
-        if (root == null) return null;
-        inorder(root, 0, 0);
-        List<List<Integer>> ans = new ArrayList<>();
-        Map<Integer, PriorityQueue<Pair>> treemap = new TreeMap<>(map);
-
-        // Iterate through map and get the result in desired format
-        for (Map.Entry<Integer, PriorityQueue<Pair>> entry : treemap.entrySet()) {
-            PriorityQueue<Pair> pq = entry.getValue();
-            // to store level along iwth nodes on that level
-            // grouping the nodes based on level
-            Map<Integer, List<Integer>> hm = new TreeMap<>();
-            List<Integer> colList = new ArrayList<>();
-
-            while (!pq.isEmpty()) {
-                Pair res = pq.poll();
-                if (hm.containsKey(res.key)) {
-                    List<Integer> list = new ArrayList<>();
-                    List<Integer> l = hm.get(res.key);
-                    l.add(res.value);
-                    hm.put(res.key, l);
-                } else hm.put(res.key, new ArrayList<>(Arrays.asList(res.value)));
+            public Pair(int key, int value) //Constructor of the class
+            {
+                this.key = key;
+                this.value = value;
             }
 
-            for (Map.Entry<Integer, List<Integer>> e : hm.entrySet()) {
-                List<Integer> pd = e.getValue();
-                Collections.sort(pd);
-                for (Integer data : pd) colList.add(data);
-            }
-            ans.add(colList);
-        }
-        return ans;
-    }
-
-    private void inorder(TreeNode root, int level, int vline) {
-        if (root == null) return;
-
-        vline -= 1;
-        inorder(root.left, level + 1, vline);
-        vline += 1;
-        // node on this vline is already present then add new node in list of this vline
-        if (map.get(vline) != null) {
-            PriorityQueue<Pair> existingpq = map.get(vline);
-            existingpq.add(new Pair(level, root.val));
-            map.put(vline, existingpq);
-        } else {
-            // Else if node on this vline is found first time then simply add it
-            PriorityQueue<Pair> newNode = new PriorityQueue<>();
-            newNode.add(new Pair(level, root.val));
-            map.put(vline, newNode);
-        }
-        inorder(root.right, level + 1, vline + 1);
-    }
-}
-
-class CanPartitionKSubsets {
-    int[] nums;
-    int n;
-    boolean[] vis;
-
-    public boolean canPartitionKSubsets(int[] nums, int k) {
-        this.n = nums.length;
-        this.nums = nums;
-        vis = new boolean[n];
-        int t = 0;
-        int max = Integer.MIN_VALUE;
-        for (int i = 0; i < n; i++) {
-            t += nums[i];
-            max = Math.max(max, nums[i]);
-        }
-
-        if (t % k != 0) return false;
-
-        for (int i = 0; i < n; i++) if (nums[i] > max) return false;
-
-        int target = t / k;
-        return dfs(0, target, 0, k);
-    }
-
-    private boolean dfs(int sum, int target, int idx, int left) {
-        if (left == 1) return true;
-
-        if (sum == target) {
-            return dfs(0, target, 0, left - 1);
-        }
-
-        if (sum > target || idx >= n) return false;
-
-        for (int i = idx; i < n; i++) {
-            if (!vis[i]) {
-                vis[i] = true;
-                if (dfs(sum + nums[i], target, idx + 1, left)) return true;
-                vis[i] = false;
+            public int compareTo(Pair o) {
+                return this.key - o.key;
             }
         }
-        return false;
+
+        // To store the vline and nodes with level on that vline
+        HashMap<Integer, PriorityQueue<Pair>> map = new HashMap<>();
+
+        public List<List<Integer>> verticalTraversal(TreeNode root) {
+            if (root == null) return null;
+            inorder(root, 0, 0);
+            List<List<Integer>> ans = new ArrayList<>();
+            Map<Integer, PriorityQueue<Pair>> treemap = new TreeMap<>(map);
+
+            // Iterate through map and get the result in desired format
+            for (Map.Entry<Integer, PriorityQueue<Pair>> entry : treemap.entrySet()) {
+                PriorityQueue<Pair> pq = entry.getValue();
+                // to store level along iwth nodes on that level
+                // grouping the nodes based on level
+                Map<Integer, List<Integer>> hm = new TreeMap<>();
+                List<Integer> colList = new ArrayList<>();
+
+                while (!pq.isEmpty()) {
+                    Pair res = pq.poll();
+                    if (hm.containsKey(res.key)) {
+                        List<Integer> list = new ArrayList<>();
+                        List<Integer> l = hm.get(res.key);
+                        l.add(res.value);
+                        hm.put(res.key, l);
+                    } else hm.put(res.key, new ArrayList<>(Arrays.asList(res.value)));
+                }
+
+                for (Map.Entry<Integer, List<Integer>> e : hm.entrySet()) {
+                    List<Integer> pd = e.getValue();
+                    Collections.sort(pd);
+                    for (Integer data : pd) colList.add(data);
+                }
+                ans.add(colList);
+            }
+            return ans;
+        }
+
+        private void inorder(TreeNode root, int level, int vline) {
+            if (root == null) return;
+
+            vline -= 1;
+            inorder(root.left, level + 1, vline);
+            vline += 1;
+            // node on this vline is already present then add new node in list of this vline
+            if (map.get(vline) != null) {
+                PriorityQueue<Pair> existingpq = map.get(vline);
+                existingpq.add(new Pair(level, root.val));
+                map.put(vline, existingpq);
+            } else {
+                // Else if node on this vline is found first time then simply add it
+                PriorityQueue<Pair> newNode = new PriorityQueue<>();
+                newNode.add(new Pair(level, root.val));
+                map.put(vline, newNode);
+            }
+            inorder(root.right, level + 1, vline + 1);
+        }
     }
-}
+
+    class CanPartitionKSubsets {
+        int[] nums;
+        int n;
+        boolean[] vis;
+
+        public boolean canPartitionKSubsets(int[] nums, int k) {
+            this.n = nums.length;
+            this.nums = nums;
+            vis = new boolean[n];
+            int t = 0;
+            int max = Integer.MIN_VALUE;
+            for (int i = 0; i < n; i++) {
+                t += nums[i];
+                max = Math.max(max, nums[i]);
+            }
+
+            if (t % k != 0) return false;
+
+            for (int i = 0; i < n; i++) if (nums[i] > max) return false;
+
+            int target = t / k;
+            return dfs(0, target, 0, k);
+        }
+
+        private boolean dfs(int sum, int target, int idx, int left) {
+            if (left == 1) return true;
+
+            if (sum == target) {
+                return dfs(0, target, 0, left - 1);
+            }
+
+            if (sum > target || idx >= n) return false;
+
+            for (int i = idx; i < n; i++) {
+                if (!vis[i]) {
+                    vis[i] = true;
+                    if (dfs(sum + nums[i], target, idx + 1, left)) return true;
+                    vis[i] = false;
+                }
+            }
+            return false;
+        }
+    }
 
 
 // Intial thoughts:-
 // BFS algorithm
 
 
-//  DP
+    //  DP
 //  - create a map to adjancent nodes alogn with distance
 //  - iterate through all nodes and calculate cost of each path recursively
 //  - update minCost path if currNode reaches to destination
 //  - returm minCost;
 // TC  = (2^n)
 // SC = O(n)
-class FindCheapestPriceGraphBFS {
+    class FindCheapestPriceGraphBFS {
 
-    /*
-    class Solution {
-  int n;
-  int[][] flights;
-  int src;
-  int dst;
-  // create a graph
-  Map<Integer, List < pair >> graph;
-  int minCost = Integer.MAX_VALUE;
-  public int findCheapestPrice(int n, int[][] flights, int src, int dst, int k) {
-    this.n = n;
-    this.flights = flights;
-    this.src = src;
-    this.dst = dst;
-    this.graph = new HashMap<>();
-    //Build graph used for adjancy list and distance
-    for (int i = 0; i < flights.length; i++) {
-      int srcF = flights[i][0];
-      int dest = flights[i][1];
-      int dist = flights[i][2];
-      if (graph.containsKey(srcF)) {
-        List < pair > existing = graph.get(srcF);
-        existing.add(new pair(dest, dist));
-        graph.put(srcF, existing);
-      } else {
-        List < pair > newNodes = new ArrayList < > ();
-        newNodes.add(new pair(dest, dist));
-        graph.put(srcF, newNodes);
-       }
-    }
+        /*
+        class Solution {
+      int n;
+      int[][] flights;
+      int src;
+      int dst;
+      // create a graph
+      Map<Integer, List < pair >> graph;
+      int minCost = Integer.MAX_VALUE;
+      public int findCheapestPrice(int n, int[][] flights, int src, int dst, int k) {
+        this.n = n;
+        this.flights = flights;
+        this.src = src;
+        this.dst = dst;
+        this.graph = new HashMap<>();
+        //Build graph used for adjancy list and distance
+        for (int i = 0; i < flights.length; i++) {
+          int srcF = flights[i][0];
+          int dest = flights[i][1];
+          int dist = flights[i][2];
+          if (graph.containsKey(srcF)) {
+            List < pair > existing = graph.get(srcF);
+            existing.add(new pair(dest, dist));
+            graph.put(srcF, existing);
+          } else {
+            List < pair > newNodes = new ArrayList < > ();
+            newNodes.add(new pair(dest, dist));
+            graph.put(srcF, newNodes);
+           }
+        }
 
-    int  result  =  cheapest(0, k, src) ;
-    return result == Integer.MAX_VALUE ? -1 : res;
-  }
+        int  result  =  cheapest(0, k, src) ;
+        return result == Integer.MAX_VALUE ? -1 : res;
+      }
 
-  private int cheapest(int currStop, int maxStops, int currNode) {
-    // base case
+      private int cheapest(int currStop, int maxStops, int currNode) {
+        // base case
 
-    if (maxStops == 0) {
-     for (pair p :  graph.get(currNode)){
-          if (p.node == dst) return p.dist;
+        if (maxStops == 0) {
+         for (pair p :  graph.get(currNode)){
+              if (p.node == dst) return p.dist;
+          }
+        }
+        if (currNode == dst) return 0;
+        if (currStop == maxStops) return -1;
+
+        int currCost = 0;
+        for (pair p :  graph.get(currNode)){
+            int price = cheapest(currStop + 1, maxStops, p.node);
+            if  (price != -1){
+              currCost  = p.dist + price;
+              currCost += p.dist; // add last edge cost
+              minCost = Math.min(minCost, currCost);
+            }
+        }
+        return minCost;
+      }
+
+      class pair {
+        int node;
+        int dist;
+        pair(int node, int dist) {
+          this.node = node;
+          this.dist = dist;
+        }
+
       }
     }
-    if (currNode == dst) return 0;
-    if (currStop == maxStops) return -1;
+         */
+        Map<Integer, List<int[]>> map;
+        Integer dp[][];
 
-    int currCost = 0;
-    for (pair p :  graph.get(currNode)){
-        int price = cheapest(currStop + 1, maxStops, p.node);
-        if  (price != -1){
-          currCost  = p.dist + price;
-          currCost += p.dist; // add last edge cost
-          minCost = Math.min(minCost, currCost);
-        }
-    }
-    return minCost;
-  }
-
-  class pair {
-    int node;
-    int dist;
-    pair(int node, int dist) {
-      this.node = node;
-      this.dist = dist;
-    }
-
-  }
-}
-     */
-    Map<Integer, List<int[]>> map;
-    Integer dp[][];
-
-    int find(int src, int dest, int k) {
-        if (k < 0) return 1000_000_00;
-        if (src == dest) return 0;
-        if (dp[src][k] != null) return dp[src][k];
-        int max = 1000_000_00;
-        for (int arr[] : map.getOrDefault(src, new ArrayList<>())) {
-            max = Math.min(max, arr[1] + find(arr[0], dest, k - 1));
-        }
-        return dp[src][k] = max;
-    }
-
-    public int findCheapestPrice(int n, int[][] flights, int src, int dst, int K) {
-        map = new HashMap<>();
-        dp = new Integer[n + 1][K + 2];
-        for (int a[] : flights) {
-            map.computeIfAbsent(a[0], k -> new ArrayList<>());
-            map.get(a[0]).add(new int[]{a[1], a[2]});
-        }
-        int temp = find(src, dst, K + 1);
-        return temp >= 1000_000_00 ? -1 : temp;
-    }
-}
-
-// TODO: Solve it again
-class WaysToArriveAtDestnation {
-    public int countPaths(int n, int[][] roads) {
-        final List<List<Node>> graph = new ArrayList<>(n);
-        for (int i = 0; i < n; i++) {
-            graph.add(new ArrayList());
-        }
-        for (final int[] arr : roads) {
-            graph.get(arr[0]).add(new Node(arr[1], arr[2]));
-            graph.get(arr[1]).add(new Node(arr[0], arr[2]));
-        }
-        return this.dfs(graph, n);
-    }
-
-    public int dfs(final List<List<Node>> adj, int n) {
-        final int mod = 1_000_000_007;
-        final Queue<Node> queue = new PriorityQueue<>(n);
-        final long[] costs = new long[n];
-        final long[] ways = new long[n];
-        final boolean[] cache = new boolean[n];
-        queue.add(new Node(0, 0));
-        Arrays.fill(costs, Long.MAX_VALUE);
-        costs[0] = 0;
-        //one way to visit first node
-        ways[0] = 1;
-        while (!queue.isEmpty()) {
-            final Node currentNode = queue.poll();
-            if (currentNode.cost > costs[currentNode.position] || cache[currentNode.position]) {
-                continue;
+        int find(int src, int dest, int k) {
+            if (k < 0) return 1000_000_00;
+            if (src == dest) return 0;
+            if (dp[src][k] != null) return dp[src][k];
+            int max = 1000_000_00;
+            for (int arr[] : map.getOrDefault(src, new ArrayList<>())) {
+                max = Math.min(max, arr[1] + find(arr[0], dest, k - 1));
             }
-            for (final Node vertex : adj.get(currentNode.position)) {
-                if (costs[currentNode.position] + vertex.cost < costs[vertex.position]) {
-                    costs[vertex.position] = costs[currentNode.position] + vertex.cost;
-                    ways[vertex.position] = ways[currentNode.position] % mod;
-                    queue.add(new Node(vertex.position, costs[vertex.position]));
-                } else if (costs[currentNode.position] + vertex.cost == costs[vertex.position]) {
-                    ways[vertex.position] = (ways[vertex.position] + ways[currentNode.position]) % mod;
+            return dp[src][k] = max;
+        }
+
+        public int findCheapestPrice(int n, int[][] flights, int src, int dst, int K) {
+            map = new HashMap<>();
+            dp = new Integer[n + 1][K + 2];
+            for (int a[] : flights) {
+                map.computeIfAbsent(a[0], k -> new ArrayList<>());
+                map.get(a[0]).add(new int[]{a[1], a[2]});
+            }
+            int temp = find(src, dst, K + 1);
+            return temp >= 1000_000_00 ? -1 : temp;
+        }
+    }
+
+    // TODO: Solve it again
+    class WaysToArriveAtDestnation {
+        public int countPaths(int n, int[][] roads) {
+            final List<List<Node>> graph = new ArrayList<>(n);
+            for (int i = 0; i < n; i++) {
+                graph.add(new ArrayList());
+            }
+            for (final int[] arr : roads) {
+                graph.get(arr[0]).add(new Node(arr[1], arr[2]));
+                graph.get(arr[1]).add(new Node(arr[0], arr[2]));
+            }
+            return this.dfs(graph, n);
+        }
+
+        public int dfs(final List<List<Node>> adj, int n) {
+            final int mod = 1_000_000_007;
+            final Queue<Node> queue = new PriorityQueue<>(n);
+            final long[] costs = new long[n];
+            final long[] ways = new long[n];
+            final boolean[] cache = new boolean[n];
+            queue.add(new Node(0, 0));
+            Arrays.fill(costs, Long.MAX_VALUE);
+            costs[0] = 0;
+            //one way to visit first node
+            ways[0] = 1;
+            while (!queue.isEmpty()) {
+                final Node currentNode = queue.poll();
+                if (currentNode.cost > costs[currentNode.position] || cache[currentNode.position]) {
+                    continue;
                 }
-            }
-        }
-        return (int) ways[n - 1];
-    }
-
-    @SuppressWarnings("ALL")
-    private static class Node implements Comparable<Node> {
-        int position;
-        long cost;
-
-        public Node(int dis, long val) {
-            this.position = dis;
-            this.cost = val;
-        }
-
-
-        @Override
-        public int compareTo(final Node o) {
-            return Long.compare(this.cost, o.cost);
-        }
-    }
-}
-
-
-// TODO: Revisit
-class MaxMatrixSum {
-    public long maxMatrixSum(int[][] matrix) {
-        int m = matrix.length;
-        int n = matrix[0].length;
-        int mini = Integer.MAX_VALUE;
-        int cnt = 0;
-        long sum = 0;
-        for (int i = 0; i < m; i++) {
-            for (int j = 0; j < n; j++) {
-                sum += Math.abs(matrix[i][j]);
-                if (matrix[i][j] < 0) cnt++;
-                mini = Math.min(mini, Math.abs(matrix[i][j]));
-            }
-        }
-        if (cnt % 2 == 0) return sum;
-        return sum - 2 * mini;
-    }
-}
-
-class WordladderLengthBFS {
-    public int ladderLength(String start, String end, List<String> wordList) {
-        Set<String> dict = new HashSet<>(wordList);
-
-        dict.add(start);
-        Map<String, Integer> distances = new HashMap<>();
-        return bfsHelper(start, end, dict, distances);
-
-    }
-
-    // use BFS, and traverse each word and its neighbors
-    // if we find any word equals to end, we've return the "level" -- distances -- length
-    private int bfsHelper(String start,
-                          String end,
-                          Set<String> dict,
-                          Map<String, Integer> distances) {
-        Queue<String> queue = new LinkedList<>();
-        queue.offer(start);
-        distances.put(start, 0); // distance keep track of how far it is from start
-
-        while (!queue.isEmpty()) {
-            String word = queue.poll();
-            int distance = distances.get(word);
-
-            for (String nextWord : getNextWords(word, dict)) {
-                if (!distances.containsKey(nextWord)) {
-                    queue.offer(nextWord);
-                    distances.put(nextWord, distance + 1);
-                }
-            }
-        }
-
-        printMap(distances);
-
-        if (distances.containsKey(end)) {
-            return distances.get(end) + 1;
-        }
-        return 0;
-    }
-
-    private void printMap(Map<String, Integer> visited) {
-        for (String s : visited.keySet()) {
-            System.out.println(s + " : " + visited.get(s));
-        }
-    }
-
-    private List<String> getNextWords(String word, Set<String> dict) {
-        List<String> res = new ArrayList<>();
-        for (int i = 0; i < word.length(); i++) {
-            char ch = word.charAt(i);
-            for (int j = 0; j < 26; j++) {
-                if (ch != (char) ('a' + j)) {
-                    // substitute word.charAt(i) using (char)('a' + j)
-                    String nextWord = word.substring(0, i) + (char) ('a' + j) + word.substring(i + 1);
-                    if (dict.contains(nextWord)) {
-                        res.add(nextWord);
+                for (final Node vertex : adj.get(currentNode.position)) {
+                    if (costs[currentNode.position] + vertex.cost < costs[vertex.position]) {
+                        costs[vertex.position] = costs[currentNode.position] + vertex.cost;
+                        ways[vertex.position] = ways[currentNode.position] % mod;
+                        queue.add(new Node(vertex.position, costs[vertex.position]));
+                    } else if (costs[currentNode.position] + vertex.cost == costs[vertex.position]) {
+                        ways[vertex.position] = (ways[vertex.position] + ways[currentNode.position]) % mod;
                     }
                 }
             }
+            return (int) ways[n - 1];
         }
-        return res;
-    }
-}
 
-class BoundaryPathsTree {
+        @SuppressWarnings("ALL")
+        private  class Node implements Comparable<Node> {
+            int position;
+            long cost;
 
-    int[] dx = {0, 1, 0, -1};
-    int[] dy = {1, 0, -1, 0};
-    int M = 1000000000 + 7;
-    int m, n;
-
-    public int findPaths(int m, int n, int maxMove, int startRow, int startColumn) {
-        this.m = m;
-        this.n = n;
-        int[][][] dp = new int[m][n][maxMove + 1];
-        for (int[][] l : dp)
-            for (int[] sl : l) Arrays.fill(sl, -1);
-
-        return recursive(m, n, startRow, startColumn, dp, maxMove);
-    }
-
-    // this function will tell that the current node is safe to be visited or not
-    private boolean isSafe(int r, int c) {
-        return (r < m && c < n && r >= 0 && c >= 0);
-    }
-
-    private int recursive(int m, int n, int r, int c, int[][][] dp, int maxMove) {
-        // base cases
-        if (!isSafe(r, c)) return 1;
-        if (maxMove == 0) return 0;
-
-        if (dp[r][c][maxMove] >= 0) return dp[r][c][maxMove];
-
-        dp[r][c][maxMove] = (
-                (recursive(m, n, r + 1, c, dp, maxMove - 1)
-                        + recursive(m, n, r - 1, c, dp, maxMove - 1)) % M
-                        + (recursive(m, n, r, c + 1, dp, maxMove - 1)
-                        + recursive(m, n, r, c - 1, dp, maxMove - 1)) % M
-        ) % M;
-
-        return dp[r][c][maxMove];
-    }
-}
+            public Node(int dis, long val) {
+                this.position = dis;
+                this.cost = val;
+            }
 
 
-//TODO: Revisit
-
-class UniquePathsIIIBFSGrid {
-    int[] dx = {0, 1, 0, -1};
-    int[] dy = {1, 0, -1, 0};
-    boolean[][] visited;
-    int[][] obstacleGrid;
-    int m, n, space = 0;
-    static int paths = 0;
-    // find src and dest points
-    Point src, dest;
-
-    public int uniquePathsIII(int[][] obstacleGrid) {
-        m = obstacleGrid.length;
-        n = obstacleGrid[0].length;
-
-        if (m == 1 && n == 1) {
-            if (obstacleGrid[0][0] == 0) return 1;
-            else return 0;
+            @Override
+            public int compareTo(final Node o) {
+                return Long.compare(this.cost, o.cost);
+            }
         }
-        this.src = new Point();
-        this.dest = new Point();
+    }
 
-        for (int i = 0; i < m; i++) {
-            for (int j = 0; j < n; j++) {
-                if (obstacleGrid[i][j] == 1) {
-                    this.src.x = i;
-                    this.src.y = j;
-                } else if (obstacleGrid[i][j] == 2) {
-                    this.dest.x = i;
-                    this.dest.y = j;
-                } else if (obstacleGrid[i][j] == 0) {
-                    space++;
+
+    // TODO: Revisit
+    class MaxMatrixSum {
+        public long maxMatrixSum(int[][] matrix) {
+            int m = matrix.length;
+            int n = matrix[0].length;
+            int mini = Integer.MAX_VALUE;
+            int cnt = 0;
+            long sum = 0;
+            for (int i = 0; i < m; i++) {
+                for (int j = 0; j < n; j++) {
+                    sum += Math.abs(matrix[i][j]);
+                    if (matrix[i][j] < 0) cnt++;
+                    mini = Math.min(mini, Math.abs(matrix[i][j]));
                 }
             }
+            if (cnt % 2 == 0) return sum;
+            return sum - 2 * mini;
+        }
+    }
+
+    class WordladderLengthBFS {
+        public int ladderLength(String start, String end, List<String> wordList) {
+            Set<String> dict = new HashSet<>(wordList);
+
+            dict.add(start);
+            Map<String, Integer> distances = new HashMap<>();
+            return bfsHelper(start, end, dict, distances);
+
         }
 
-        this.obstacleGrid = obstacleGrid;
-        visited = new boolean[m][n];
-        int[][] dp = new int[m][n];
-        for (int[] rows : dp) Arrays.fill(rows, -1);
-        recursive(m, n, this.src.x, this.src.y, 0);
-        // For a single path once found paths=0 is done to reset the path again and evaluate it
-        int k = paths;
-        paths = 0;
-        return k;
-    }
+        // use BFS, and traverse each word and its neighbors
+        // if we find any word equals to end, we've return the "level" -- distances -- length
+        private int bfsHelper(String start,
+                              String end,
+                              Set<String> dict,
+                              Map<String, Integer> distances) {
+            Queue<String> queue = new LinkedList<>();
+            queue.offer(start);
+            distances.put(start, 0); // distance keep track of how far it is from start
 
-    // this function will tell that the current node is safe to be visited or not
-    private boolean isSafe(int r, int c) {
-        return (r < m && c < n && r >= 0 && c >= 0 && (obstacleGrid[r][c] != -1));
-    }
+            while (!queue.isEmpty()) {
+                String word = queue.poll();
+                int distance = distances.get(word);
 
-    private void recursive(int m, int n, int r, int c, int cells) {
-        // base cases
-        if (r == this.dest.x && c == this.dest.y && cells == space) {
-            paths++;
-            return;
-        } else if (!isSafe(r, c) || visited[r][c]) return;
-        for (int i = 0; i < 4; i++) {
-            int x = r + dx[i];
-            int y = c + dy[i];
-            if (isSafe(x, y) && !visited[x][y]) {
-                visited[r][c] = true;
-                if (obstacleGrid[x][y] == 0) cells += 1;
-                recursive(m, n, x, y, cells);
-                if (obstacleGrid[x][y] == 0) cells -= 1;
-                visited[r][c] = false; // backtrack
+                for (String nextWord : getNextWords(word, dict)) {
+                    if (!distances.containsKey(nextWord)) {
+                        queue.offer(nextWord);
+                        distances.put(nextWord, distance + 1);
+                    }
+                }
+            }
+
+            printMap(distances);
+
+            if (distances.containsKey(end)) {
+                return distances.get(end) + 1;
+            }
+            return 0;
+        }
+
+        private void printMap(Map<String, Integer> visited) {
+            for (String s : visited.keySet()) {
+                System.out.println(s + " : " + visited.get(s));
             }
         }
-        return;
+
+        private List<String> getNextWords(String word, Set<String> dict) {
+            List<String> res = new ArrayList<>();
+            for (int i = 0; i < word.length(); i++) {
+                char ch = word.charAt(i);
+                for (int j = 0; j < 26; j++) {
+                    if (ch != (char) ('a' + j)) {
+                        // substitute word.charAt(i) using (char)('a' + j)
+                        String nextWord = word.substring(0, i) + (char) ('a' + j) + word.substring(i + 1);
+                        if (dict.contains(nextWord)) {
+                            res.add(nextWord);
+                        }
+                    }
+                }
+            }
+            return res;
+        }
     }
 
-    class Point {
-        int x;
-        int y;
+    class BoundaryPathsTree {
 
-        Point(int x, int y) {
-            this.x = x;
-            this.y = y;
+        int[] dx = {0, 1, 0, -1};
+        int[] dy = {1, 0, -1, 0};
+        int M = 1000000000 + 7;
+        int m, n;
+
+        public int findPaths(int m, int n, int maxMove, int startRow, int startColumn) {
+            this.m = m;
+            this.n = n;
+            int[][][] dp = new int[m][n][maxMove + 1];
+            for (int[][] l : dp)
+                for (int[] sl : l) Arrays.fill(sl, -1);
+
+            return recursive(m, n, startRow, startColumn, dp, maxMove);
         }
 
-        Point() {
+        // this function will tell that the current node is safe to be visited or not
+        private boolean isSafe(int r, int c) {
+            return (r < m && c < n && r >= 0 && c >= 0);
+        }
+
+        private int recursive(int m, int n, int r, int c, int[][][] dp, int maxMove) {
+            // base cases
+            if (!isSafe(r, c)) return 1;
+            if (maxMove == 0) return 0;
+
+            if (dp[r][c][maxMove] >= 0) return dp[r][c][maxMove];
+
+            dp[r][c][maxMove] = (
+                    (recursive(m, n, r + 1, c, dp, maxMove - 1)
+                            + recursive(m, n, r - 1, c, dp, maxMove - 1)) % M
+                            + (recursive(m, n, r, c + 1, dp, maxMove - 1)
+                            + recursive(m, n, r, c - 1, dp, maxMove - 1)) % M
+            ) % M;
+
+            return dp[r][c][maxMove];
         }
     }
-}
+
 
 //TODO: Revisit
-class SQRTBinarySeach {
-    public int mySqrt(int num) {
 
-        if (num < 2) {
-            return num;
+    class UniquePathsIIIBFSGrid {
+        int[] dx = {0, 1, 0, -1};
+        int[] dy = {1, 0, -1, 0};
+        boolean[][] visited;
+        int[][] obstacleGrid;
+        int m, n, space = 0;
+        int paths = 0;
+        // find src and dest points
+        Point src, dest;
+
+        public int uniquePathsIII(int[][] obstacleGrid) {
+            m = obstacleGrid.length;
+            n = obstacleGrid[0].length;
+
+            if (m == 1 && n == 1) {
+                if (obstacleGrid[0][0] == 0) return 1;
+                else return 0;
+            }
+            this.src = new Point();
+            this.dest = new Point();
+
+            for (int i = 0; i < m; i++) {
+                for (int j = 0; j < n; j++) {
+                    if (obstacleGrid[i][j] == 1) {
+                        this.src.x = i;
+                        this.src.y = j;
+                    } else if (obstacleGrid[i][j] == 2) {
+                        this.dest.x = i;
+                        this.dest.y = j;
+                    } else if (obstacleGrid[i][j] == 0) {
+                        space++;
+                    }
+                }
+            }
+
+            this.obstacleGrid = obstacleGrid;
+            visited = new boolean[m][n];
+            int[][] dp = new int[m][n];
+            for (int[] rows : dp) Arrays.fill(rows, -1);
+            recursive(m, n, this.src.x, this.src.y, 0);
+            // For a single path once found paths=0 is done to reset the path again and evaluate it
+            int k = paths;
+            paths = 0;
+            return k;
         }
 
-        long start = 2;
-        long end = num;
-        while (start <= end) {
-            long mid = start + (end - start) / 2;
-            if (mid * mid == num) return (int) mid;
-            else if (mid * mid < num) start = mid + 1;
-            else end = mid - 1;
+        // this function will tell that the current node is safe to be visited or not
+        private boolean isSafe(int r, int c) {
+            return (r < m && c < n && r >= 0 && c >= 0 && (obstacleGrid[r][c] != -1));
         }
-        return (int) (start - 1);
+
+        private void recursive(int m, int n, int r, int c, int cells) {
+            // base cases
+            if (r == this.dest.x && c == this.dest.y && cells == space) {
+                paths++;
+                return;
+            } else if (!isSafe(r, c) || visited[r][c]) return;
+            for (int i = 0; i < 4; i++) {
+                int x = r + dx[i];
+                int y = c + dy[i];
+                if (isSafe(x, y) && !visited[x][y]) {
+                    visited[r][c] = true;
+                    if (obstacleGrid[x][y] == 0) cells += 1;
+                    recursive(m, n, x, y, cells);
+                    if (obstacleGrid[x][y] == 0) cells -= 1;
+                    visited[r][c] = false; // backtrack
+                }
+            }
+            return;
+        }
+
+        class Point {
+            int x;
+            int y;
+
+            Point(int x, int y) {
+                this.x = x;
+                this.y = y;
+            }
+
+            Point() {
+            }
+        }
     }
 
-    // O(n+m), O(1)
+    //TODO: Revisit
+    class SQRTBinarySeach {
+        public int mySqrt(int num) {
+
+            if (num < 2) {
+                return num;
+            }
+
+            long start = 2;
+            long end = num;
+            while (start <= end) {
+                long mid = start + (end - start) / 2;
+                if (mid * mid == num) return (int) mid;
+                else if (mid * mid < num) start = mid + 1;
+                else end = mid - 1;
+            }
+            return (int) (start - 1);
+        }
+
+        // O(n+m), O(1)
 	/*
      This problem is called Merge Sorted Array
      Example:-
@@ -10803,284 +10803,285 @@ Output: [1,2,2,3,5,6]
         After that remaining element of nums2 is utitlised
         */
 
-    public void merge(int[] nums1, int m, int[] nums2, int n) {
-        int i = m - 1, j = n - 1, k = m + n - 1;
-        while (j > -1) {
-            if (i > -1) {
-                nums1[k--] = nums1[i] > nums2[j] ? nums1[i--] : nums2[j--];
-            } else {
-                nums1[k--] = nums2[j--];
-            }
-        }
-    }
-
-
-    // This is based on binary search of peak element
-    // Input: nums = [1,2,3,1]
-    // Output: 2
-    // Explanation: 3 is a peak element and your function should return the index number 2.
-    // TC = O(logn), SC = O(1)
-    public int findPeakElement(int[] nums) {
-        int n = nums.length;
-        int l = 0, r = n - 1;
-        while (l < r) {
-            int m = l + (r - l) / 2;
-            if (nums[m] < nums[m + 1]) l = m + 1;
-            else r = m;
-        }
-        return l;
-    }
-
-    // O(logn)
-    // We have to iteratively partition array into haves and evavluate which half is sorted
-    // and then apply BS for the element
-    public int search(int[] nums, int target) {
-        int n = nums.length;
-        int l = 0, r = n - 1;
-
-        // Find 2 halves of array
-        while (l <= r) {
-            int mid = l + (r - l) / 2;
-            if (nums[mid] == target) return mid;
-            // First half is sorted and other half is unsorted
-            if (nums[l] <= nums[mid]) {
-                if (target >= nums[l] && target < nums[mid]) {
-                    r = mid - 1;
-                } else l = mid + 1;
-            }
-            // Second half is sorted and other half is unsorted
-            else {
-                if (target > nums[mid] && target <= nums[r]) {
-                    l = mid + 1;
-                } else r = mid - 1;
-            }
-        }
-        return -1;
-
-    }
-
-    // PQ + BFS
-    // This is a astandard application for PQ and application of BFS
-    //[[u,v] = meeting times list]
-    Map<Integer, Map<Integer, List<Integer>>> g = new HashMap<>();
-
-    public List<Integer> findAllPeople(int n, int[][] meetings, int firstPerson) {
-
-        int[] notified = new int[n];
-        Arrays.fill(notified, -1);
-        for (int i = 0; i < n; i++) g.put(i, new HashMap<>());
-        for (int[] m : meetings) getMeetingTimes(m[0], m[1]).add(m[2]);
-        List<Integer> res = new ArrayList<>();
-
-        PriorityQueue<int[]> pq = new PriorityQueue<>((a, b) -> a[1] - b[1]);
-        pq.offer(new int[]{
-                0,
-                0
-        });
-        pq.offer(new int[]{
-                firstPerson,
-                0
-        });
-
-        while (!pq.isEmpty()) {
-            int[] cur = pq.poll();
-            int u = cur[0];
-            if (notified[u] > -1) continue; //already visited
-            notified[u] = cur[1];
-            res.add(u);
-            for (Map.Entry<Integer, List<Integer>> entry : g.get(u).entrySet()) {
-                int v = entry.getKey();
-                if (notified[v] > -1) continue; // already visited
-                Collections.sort(entry.getValue());
-                for (int time : entry.getValue()) {
-                    //get first time u and v met after person u learned the secret
-                    if (time >= notified[u]) {
-                        pq.offer(new int[]{
-                                v,
-                                time
-                        });
-                        break;
-                    }
+        public void merge(int[] nums1, int m, int[] nums2, int n) {
+            int i = m - 1, j = n - 1, k = m + n - 1;
+            while (j > -1) {
+                if (i > -1) {
+                    nums1[k--] = nums1[i] > nums2[j] ? nums1[i--] : nums2[j--];
+                } else {
+                    nums1[k--] = nums2[j--];
                 }
             }
-            g.remove(u); //u was fully processed
         }
 
-        return res;
-    }
 
-    private List<Integer> getMeetingTimes(int u, int v) {
-        List<Integer> times = g.get(u).get(v);
-        if (times == null) {
-            times = new ArrayList<>();
-            g.get(u).put(v, times);
-            g.get(v).put(u, times);
+        // This is based on binary search of peak element
+        // Input: nums = [1,2,3,1]
+        // Output: 2
+        // Explanation: 3 is a peak element and your function should return the index number 2.
+        // TC = O(logn), SC = O(1)
+        public int findPeakElement(int[] nums) {
+            int n = nums.length;
+            int l = 0, r = n - 1;
+            while (l < r) {
+                int m = l + (r - l) / 2;
+                if (nums[m] < nums[m + 1]) l = m + 1;
+                else r = m;
+            }
+            return l;
         }
-        return times;
-    }
+
+        // O(logn)
+        // We have to iteratively partition array into haves and evavluate which half is sorted
+        // and then apply BS for the element
+        public int search(int[] nums, int target) {
+            int n = nums.length;
+            int l = 0, r = n - 1;
+
+            // Find 2 halves of array
+            while (l <= r) {
+                int mid = l + (r - l) / 2;
+                if (nums[mid] == target) return mid;
+                // First half is sorted and other half is unsorted
+                if (nums[l] <= nums[mid]) {
+                    if (target >= nums[l] && target < nums[mid]) {
+                        r = mid - 1;
+                    } else l = mid + 1;
+                }
+                // Second half is sorted and other half is unsorted
+                else {
+                    if (target > nums[mid] && target <= nums[r]) {
+                        l = mid + 1;
+                    } else r = mid - 1;
+                }
+            }
+            return -1;
+
+        }
+
+        // PQ + BFS
+        // This is a astandard application for PQ and application of BFS
+        //[[u,v] = meeting times list]
+        Map<Integer, Map<Integer, List<Integer>>> g = new HashMap<>();
+
+        public List<Integer> findAllPeople(int n, int[][] meetings, int firstPerson) {
+
+            int[] notified = new int[n];
+            Arrays.fill(notified, -1);
+            for (int i = 0; i < n; i++) g.put(i, new HashMap<>());
+            for (int[] m : meetings) getMeetingTimes(m[0], m[1]).add(m[2]);
+            List<Integer> res = new ArrayList<>();
+
+            PriorityQueue<int[]> pq = new PriorityQueue<>((a, b) -> a[1] - b[1]);
+            pq.offer(new int[]{
+                    0,
+                    0
+            });
+            pq.offer(new int[]{
+                    firstPerson,
+                    0
+            });
+
+            while (!pq.isEmpty()) {
+                int[] cur = pq.poll();
+                int u = cur[0];
+                if (notified[u] > -1) continue; //already visited
+                notified[u] = cur[1];
+                res.add(u);
+                for (Map.Entry<Integer, List<Integer>> entry : g.get(u).entrySet()) {
+                    int v = entry.getKey();
+                    if (notified[v] > -1) continue; // already visited
+                    Collections.sort(entry.getValue());
+                    for (int time : entry.getValue()) {
+                        //get first time u and v met after person u learned the secret
+                        if (time >= notified[u]) {
+                            pq.offer(new int[]{
+                                    v,
+                                    time
+                            });
+                            break;
+                        }
+                    }
+                }
+                g.remove(u); //u was fully processed
+            }
+
+            return res;
+        }
+
+        private List<Integer> getMeetingTimes(int u, int v) {
+            List<Integer> times = g.get(u).get(v);
+            if (times == null) {
+                times = new ArrayList<>();
+                g.get(u).put(v, times);
+                g.get(v).put(u, times);
+            }
+            return times;
+        }
 
 
-    // TC = O(2^n), SC = O(2^n)
-    // this is based on generating all possivble permuatation of an array of desired length
+        // TC = O(2^n), SC = O(2^n)
+        // this is based on generating all possivble permuatation of an array of desired length
 //    via backtracking
-    Set<Integer> set = new HashSet<>();
+        Set<Integer> set = new HashSet<>();
 
-    public int[] findEvenNumbers(int[] digits) {
-        int n = digits.length;
-        boolean[] seen = new boolean[n];
-        permute(0, digits, "", seen);
-        return set.stream().mapToInt(x -> x).sorted().toArray();
-    }
-
-    private void permute(int idx, int[] digits, String curr, boolean[] seen) {
-
-        //base cases
-        if (idx == 3) {
-            set.add(Integer.parseInt(curr));
-            return;
+        public int[] findEvenNumbers(int[] digits) {
+            int n = digits.length;
+            boolean[] seen = new boolean[n];
+            permute(0, digits, "", seen);
+            return set.stream().mapToInt(x -> x).sorted().toArray();
         }
 
-        for (int i = 0; i < digits.length; i++) {
-            if (seen[i] || (idx == 0 && digits[i] == 0) || (idx == 2 && digits[i] % 2 != 0)) continue;
-            seen[i] = true;
-            permute(idx + 1, digits, curr + digits[i], seen);
-            seen[i] = false;
-        }
-    }
+        private void permute(int idx, int[] digits, String curr, boolean[] seen) {
 
-
-    // TC = (n/2), SC = O(1)
-    public ListNode deleteMiddle(ListNode head) {
-        if (head.next == null) return null;
-        ListNode slow = head;
-        ListNode fast = head.next.next;
-        while (fast != null && fast.next != null) {
-            slow = slow.next;
-            fast = fast.next.next;
-        }
-        slow.next = slow.next.next;
-        return head;
-    }
-
-
-    private ListNode deleteMiddleSimpleApproach(ListNode head) {
-        if (head == null || head.next == null) return null;
-        ListNode curr = head;
-        ListNode prev = null;
-        ListNode slow = head;
-        ListNode fast = head;
-        while (fast != null && fast.next != null) {
-            prev = curr;
-            slow = slow.next;
-            fast = fast.next.next;
-        }
-        prev.next = curr.next;
-        return head;
-    }
-
-
-    // LCA based solution
-    public TreeNode findLCA(TreeNode node, int s, int d) {
-        if (node == null) return null;
-
-        if (node.val == s || node.val == d) return node;
-
-        TreeNode left = findLCA(node.left, s, d);
-        TreeNode right = findLCA(node.right, s, d);
-        if (left != null && right != null) return node;
-        if (left == null && right != null) return right;
-        else return left;
-    }
-
-    public String getDirections(TreeNode root, int startValue, int destValue) {
-        TreeNode CA = findLCA(root, startValue, destValue);
-        StringBuilder ans = new StringBuilder();
-        ArrayDeque<String> q1 = new ArrayDeque<>();
-        helper(CA, startValue, q1);
-        ArrayDeque<String> q2 = new ArrayDeque<>();
-        helper(CA, destValue, q2);
-
-        for (int i = 0; i < q1.size(); i++) ans.append("U");
-        while (!q2.isEmpty()) ans.append(q2.poll());
-
-        return ans.toString();
-    }
-
-    private boolean helper(TreeNode n, int v, ArrayDeque<String> q) {
-        if (n == null) return false;
-        if (n.val == v) return true;
-
-        q.offer("L");
-        boolean left = helper(n.left, v, q);
-        if (left) return true;
-        q.removeLast();
-
-        q.offer("R");
-        boolean right = helper(n.right, v, q);
-        if (right) return true;
-        q.removeLast();
-
-        return false;
-    }
-
-    //Represent the head and tail of the singly linked list
-    ListNode curr = null;
-    ListNode tail = null;
-
-    //addNode() will add a new node to the list
-    public void addNode(int data) {
-        //Create a new node
-        ListNode newNode = new ListNode(data);
-
-        //Checks if the list is empty
-        if (curr == null) {
-            //If list is empty, both head and tail will point to new node
-            curr = newNode;
-            tail = newNode;
-        } else {
-            //newNode will be added after tail such that tail's next will point to newNode
-            tail.next = newNode;
-            //newNode will become new tail of the list
-            tail = newNode;
-        }
-    }
-
-    // Author: Anand
-    public ListNode mergeNodes(ListNode head) {
-        int sum = 0;
-        int zero = 0;
-
-        while (head != null) {
-            sum += head.val;
-            if (head.val == 0) {
-                zero++;
+            //base cases
+            if (idx == 3) {
+                set.add(Integer.parseInt(curr));
+                return;
             }
-            if (zero == 2) {
-                addNode(sum);
-                zero = 1;
-                sum = 0;
+
+            for (int i = 0; i < digits.length; i++) {
+                if (seen[i] || (idx == 0 && digits[i] == 0) || (idx == 2 && digits[i] % 2 != 0)) continue;
+                seen[i] = true;
+                permute(idx + 1, digits, curr + digits[i], seen);
+                seen[i] = false;
             }
-            head = head.next;
         }
 
 
-        return curr;
+        // TC = (n/2), SC = O(1)
+        public ListNode deleteMiddle(ListNode head) {
+            if (head.next == null) return null;
+            ListNode slow = head;
+            ListNode fast = head.next.next;
+            while (fast != null && fast.next != null) {
+                slow = slow.next;
+                fast = fast.next.next;
+            }
+            slow.next = slow.next.next;
+            return head;
+        }
+
+
+        private ListNode deleteMiddleSimpleApproach(ListNode head) {
+            if (head == null || head.next == null) return null;
+            ListNode curr = head;
+            ListNode prev = null;
+            ListNode slow = head;
+            ListNode fast = head;
+            while (fast != null && fast.next != null) {
+                prev = curr;
+                slow = slow.next;
+                fast = fast.next.next;
+            }
+            prev.next = curr.next;
+            return head;
+        }
+
+
+        // LCA based solution
+        public TreeNode findLCA(TreeNode node, int s, int d) {
+            if (node == null) return null;
+
+            if (node.val == s || node.val == d) return node;
+
+            TreeNode left = findLCA(node.left, s, d);
+            TreeNode right = findLCA(node.right, s, d);
+            if (left != null && right != null) return node;
+            if (left == null && right != null) return right;
+            else return left;
+        }
+
+        public String getDirections(TreeNode root, int startValue, int destValue) {
+            TreeNode CA = findLCA(root, startValue, destValue);
+            StringBuilder ans = new StringBuilder();
+            ArrayDeque<String> q1 = new ArrayDeque<>();
+            helper(CA, startValue, q1);
+            ArrayDeque<String> q2 = new ArrayDeque<>();
+            helper(CA, destValue, q2);
+
+            for (int i = 0; i < q1.size(); i++) ans.append("U");
+            while (!q2.isEmpty()) ans.append(q2.poll());
+
+            return ans.toString();
+        }
+
+        private boolean helper(TreeNode n, int v, ArrayDeque<String> q) {
+            if (n == null) return false;
+            if (n.val == v) return true;
+
+            q.offer("L");
+            boolean left = helper(n.left, v, q);
+            if (left) return true;
+            q.removeLast();
+
+            q.offer("R");
+            boolean right = helper(n.right, v, q);
+            if (right) return true;
+            q.removeLast();
+
+            return false;
+        }
+
+        //Represent the head and tail of the singly linked list
+        ListNode curr = null;
+        ListNode tail = null;
+
+        //addNode() will add a new node to the list
+        public void addNode(int data) {
+            //Create a new node
+            ListNode newNode = new ListNode(data);
+
+            //Checks if the list is empty
+            if (curr == null) {
+                //If list is empty, both head and tail will point to new node
+                curr = newNode;
+                tail = newNode;
+            } else {
+                //newNode will be added after tail such that tail's next will point to newNode
+                tail.next = newNode;
+                //newNode will become new tail of the list
+                tail = newNode;
+            }
+        }
+
+        // Author: Anand
+        public ListNode mergeNodes(ListNode head) {
+            int sum = 0;
+            int zero = 0;
+
+            while (head != null) {
+                sum += head.val;
+                if (head.val == 0) {
+                    zero++;
+                }
+                if (zero == 2) {
+                    addNode(sum);
+                    zero = 1;
+                    sum = 0;
+                }
+                head = head.next;
+            }
+
+
+            return curr;
+        }
     }
-}
 
-class combinationSum4BottomUpDP {
-    // TC = O(n*target), SC = O(1)
-    // Bottom up approach
-    public int combinationSum4(int[] nums, int target) {
-        int n = nums.length;
-        int[] dp = new int[target + 1];
-        dp[0] = 1; // Do not take any element at all
+    class combinationSum4BottomUpDP {
+        // TC = O(n*target), SC = O(1)
+        // Bottom up approach
+        public int combinationSum4(int[] nums, int target) {
+            int n = nums.length;
+            int[] dp = new int[target + 1];
+            dp[0] = 1; // Do not take any element at all
 
-        for (int i = 1; i <= target; i++) {
-            for (int num : nums)
-                if (i >= num) dp[i] += dp[i - num];
+            for (int i = 1; i <= target; i++) {
+                for (int num : nums)
+                    if (i >= num) dp[i] += dp[i - num];
+            }
+            return dp[target];
         }
-        return dp[target];
     }
 }
